subscriber/requirements.txt
Flask==1.1.2
paho-mqtt==1.5.1
